fiddles with the tests and runner
I think the first thing to do is really understand the specifics around the Vars and IdAttrs.

Right now I think the key is grokking the generated code for `createInitProc` so I can see what happens. Somewhere in there is the logic
to build this initial `vars` variable. I think once I grok that, I can make an equivalent set of logicks (might already have it!) and then continue with the port.